requirements.txt
torch
torchvision
torchaudio
picsellia
numpy
ultralytics
pre-commit
black
Dynaconf
PyYAML
